[BC] Now please `include` me instead of `extend`!
OK, so, when I tried to test with multiple classes and class hierarchy
that extends TinyHooks in a parent class, it turned out that there are
significant issues and I could not resolve them.
I investigated the problem and realized that with `include`
class
inheritance can be dealt with much easier.
Before I added any feature with support of multiple classes and
inheritance, I needed to change its core usage from `extend` to `include`.
Luckily, other interfaces remain the same.

Since I realized this problem during developing an important feature,
"targeting", this feature is also included in this commit.
It's unhealthy to have incompatibility changes and new features
in one commit, but I was just so exhausted to split them.